ai_search_service/requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
asyncpg
sentence-transformers
numpy
python-dotenv
pypdf
